coding/requirements.txt
mysql-connector-python==8.0.20
protobuf==3.12.2
PyYAML==5.3.1
six==1.15.0
